requirements.txt
streamlit==1.22.0
cohere==4.3.1
pdfplumber==0.9.0
python-docx==0.8.11
python-pptx==0.6.21
sqlite3==2.6.0
uuid==1.30